feat: Init prisma

+++ src/controllers/todos.ts
@@ -1,10 +1,10 @@
 import { Request, Response } from "express";
 import omit from "lodash/omit";
+import { Todo } from "@prisma/client";
 import { type UpdateTodoRequestBody } from "src/db/repository/TodoRepository";
 import * as todosRepository from "src/db/repository/TodoRepository";
 import * as userRepository from "src/db/repository/UserRepository";
 import { StatusCodes } from "http-status-codes";
-import { Todo } from "src/db/entity/Todo";
 
 const formatTodo = (todo: Todo) => ({
   ...omit(todo, "id"),
@@ -17,25 +17,19 @@ export async function createTodo(req: Request, res: Response) {
   const { description }: { description: string } = req.body;
 
   try {
-    const user = await userRepository.getUserById(Number(userId));
+    const newTodo = await todosRepository.addTodo({
+      userId: Number(userId),
+      description,
+    });
 
-    if (!user) {
-      res.sendStatus(StatusCodes.BAD_REQUEST);
-      return;
-    }
-
-    const newTodo = await todosRepository.addTodo(user, description);
-
-    const {
-      raw: [{ id }],
-    } = newTodo;
-
-    const todo = await todosRepository.getTodoById(id);
+    const todo = await todosRepository.getTodoById(newTodo.id);
 
     res.status(StatusCodes.CREATED);
 
     res.send(formatTodo(todo));
-  } catch (_err) {
+
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  } catch (err) {
     res.sendStatus(StatusCodes.INTERNAL_SERVER_ERROR);
   }
 }
@@ -50,7 +44,7 @@ export async function getTodo(req: Request, res: Response) {
     return;
   }
 
-  const todo = await todosRepository.getTodoByUuid(user, uuid);
+  const todo = await todosRepository.getTodoByUuid({ userId: user.id, uuid });
 
   if (!todo) {
     res.sendStatus(StatusCodes.NOT_FOUND);
@@ -69,7 +63,7 @@ export async function getTodos(req: Request, res: Response) {
     return;
   }
 
-  const todos = await todosRepository.getUsersTodos(user);
+  const todos = await todosRepository.getUsersTodos(user.id);
 
   res.send({
     todos: todos.map((todo) => formatTodo(todo)),
@@ -80,7 +74,7 @@ export async function deleteTodo(req: Request, res: Response) {
   const { uuid, userId } = req.params;
 
   const user = await userRepository.getUserById(Number(userId));
-  const todo = await todosRepository.getTodoByUuid(user, uuid);
+  const todo = await todosRepository.getTodoByUuid({ userId: user.id, uuid });
 
   if (!user) {
     res.sendStatus(StatusCodes.NOT_FOUND);
@@ -92,7 +86,7 @@ export async function deleteTodo(req: Request, res: Response) {
     return;
   }
 
-  await todosRepository.deleteTodo(user, uuid);
+  await todosRepository.deleteTodo({ userId: user.id, uuid });
 
   res.sendStatus(StatusCodes.NO_CONTENT);
 }
@@ -109,19 +103,26 @@ export async function updateTodo(req: Request, res: Response) {
     return;
   }
 
-  const todo = await todosRepository.getTodoByUuid(user, uuid);
+  const todo = await todosRepository.getTodoByUuid({ userId: user.id, uuid });
 
   if (!todo) {
     res.sendStatus(StatusCodes.NOT_FOUND);
     return;
   }
 
-  await todosRepository.updateTodo(user, uuid, {
-    description,
-    is_done,
+  await todosRepository.updateTodo({
+    userId: user.id,
+    uuid,
+    body: {
+      description,
+      is_done,
+    },
   });
 
-  const updatedTodo = await todosRepository.getTodoByUuid(user, uuid);
+  const updatedTodo = await todosRepository.getTodoByUuid({
+    userId: user.id,
+    uuid,
+  });
 
   res.send(formatTodo(updatedTodo));
 }
